+++ internal/storage/ledger/transactions.go
@@ -272,12 +272,12 @@ func (s *Store) CommitTransaction(ctx context.Context, tx *ledger.Transaction) e
 	}
 
 	if s.ledger.HasFeature(ledger.FeatureMovesHistory, "ON") {
-		moves := Moves{}
+		moves := ledger.Moves{}
 		postings := tx.Postings
 		slices.Reverse(postings)
 
 		for _, posting := range postings {
-			moves = append(moves, &Move{
+			moves = append(moves, &ledger.Move{
 				Ledger:            s.ledger.Name,
 				Account:           posting.Destination,
 				Amount:            (*bunpaginate.BigInt)(posting.Amount),
@@ -289,7 +289,7 @@ func (s *Store) CommitTransaction(ctx context.Context, tx *ledger.Transaction) e
 			})
 			postCommitVolumes.AddInput(posting.Destination, posting.Asset, new(big.Int).Neg(posting.Amount))
 
-			moves = append(moves, &Move{
+			moves = append(moves, &ledger.Move{
 				Ledger:            s.ledger.Name,
 				IsSource:          true,
 				Account:           posting.Source,
